CARGO: track changes in crate roots of procedural macro libraries
Changes in procedural macros should be considered as changes in the project model because the model contains paths to compiled proc macro libs, and when sources of these libs are changed, the plugin needs to recompile proc macro crates (project model reloading does it) to properly expand proc macros with new changes.

These changes add tracking for changes in crate roots of proc macro libs. Ideally, we need to track all changes including child modules, but it requires access to name resolution API which is undesirable to use here