MACRO: recover discarded PUNCT and delimiter token mappings
Real procedural macros tend to discard spans (token ids) from `TokenTree.Leaf.Punct` tokens
and subtree `Delimiter`s. This routine tries to recover them using a simple heuristic:
if a token without a mapping (with id == -1) follows a token *with* a mapping, then it most likely
should be mapped with the next token in the source. The only thing should be checked is that
its type and text is equal to the token in the source.