Support quick fixes in attribute macros
Most quick fixes now work in attribute macros, except these:
- RenameFix
- AddTurbofishFix

The implementation is based on intention actions in macros - see bec931fd505c1ecea4cf886d0324d40e94b5b1b3.

The major difference is that a quick fix will be available (and will be shown in the list of quick fixes) even if the fix actually can't be applied.
This is done in order to deliver the feature faster and because I don't expect many cases where a fix can't be applied in an attribute macro